fix: simplify issuer to provider lookup
